Drop redundancies
